init: force /system to mount as R/W
With Android 6.0 seems like we can't remount as R/W anymore if the ro flag is passed to the fs_mgr while mounting the fstab, therefore simply editing it to rw does the trick.
/system stays as R/W until it's not remounted as read-only, this is a security risk, therefore on boot completed remount it as read-only.

After this patch, if is needed to modify something in the /system partition just issue mount -o rw,remount /system as for prior Android versions.