libril: Encode the Number of MNC Digits in CellIdentity
Legacy RIL uses an integer to encode the number of MNC digits. Because the size is not fixed, leading zeroes result in ambiguity in the length of the mnc. This change adds support for passing the number of encoded digits in the most-significant nibble of the mnc integer (which is only 10 bits). Thus, on any implementation that is 16-bits or wider, the mnc info will be properly encoded and decoded with the correctly-sized string.