jf-common: healthd: Add CAP_WAKE_ALARM to service
health/2.0/default/healthd_common.cpp wants to use
timerfd_create() to register a wakealarm.

To use the timerfd_create() syscall with CLOCK_BOOTTIME_ALARM,
CAP_WAKE_ALARM is needed.

Update the README to reflect the needed cap in the init file,
see also https://r.android.com/861532

Since the v1.0 HAL is in passthrough mode, the cap is needed for the 1.0
service as well.

Change-Id: I0e71da125eeef1360b5fd8bc7a69285e0b786511
Signed-off-by: Felix <google@ix5.org>